Make a prompt step for container name
